LPS-110452 Moving to the right package
