+++ modules/apps/depot/depot-web/src/main/java/com/liferay/depot/web/internal/roles/admin/group/type/contributor/DepotEntryPermission.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.depot.web.internal.resource;
+package com.liferay.depot.web.internal.roles.admin.group.type.contributor;
 
 import com.liferay.depot.model.DepotEntry;
 import com.liferay.portal.kernel.exception.PortalException;
