+++ modules/apps/redirect/redirect-web/src/main/java/com/liferay/redirect/web/internal/display/context/RedirectDisplayContext.java
@@ -45,8 +45,8 @@ import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.redirect.model.RedirectEntry;
 import com.liferay.redirect.service.RedirectEntryLocalServiceUtil;
 import com.liferay.redirect.service.RedirectEntryServiceUtil;
-import com.liferay.redirect.web.internal.resource.RedirectEntryPermission;
 import com.liferay.redirect.web.internal.search.RedirectEntrySearch;
+import com.liferay.redirect.web.internal.security.permission.resource.RedirectEntryPermission;
 import com.liferay.redirect.web.internal.util.comparator.RedirectEntryCreateDateComparator;
 import com.liferay.redirect.web.internal.util.comparator.RedirectEntryDestinationURLComparator;
 import com.liferay.redirect.web.internal.util.comparator.RedirectEntryModifiedDateComparator;
